Now supports block size in g-code
BLOCK, START and TOOL keyboard are processed when in parenthesis.
BLOCK defines the size of the piece to be machined
START defines the position of the block relative to the tool when the
program is to start.
TOOL defines the radius of the tool as X. Tool height to be set in Z (to
do).
Create a "project repository" where G-Code programs can be stored.
Improved the 3D viewer so that the edges of the block are represented.
Artifact when tool crosses the edges is known issue.