Improvements of move timings
Removed query of status during a continuous move (e,.g. G2) to avoid
stuttering.
Improved the calculation of the movement delay to include the decoding
of the serial command and sending the acknowledge in the delay of the
very first step.
Reduced the max speed of the G0 command to get a smoother acceleration /
deceleration and less vibrations.
Fixed bug in status query when sending manual move from the "simple
shape" commands.