Move optimization, manual mode improv, bugfix
Optimized processing for rapid motion to avoid divisions and
multiplications. Fixed error in linear motion speed. Simplified move
routine to no longer need the time since the start of the movement.
Added acceleration and automatic axis change from Z back to X/Y when
using the L/R buttons.
Fixed bug in calculation of the duration of the commands in the pipe.
The circular buffer used to store each pending command duration was too
small and could overflow in case of too many short commands.