Fixed socket communication bug
Buffer sent over TCP/IP get concatenated and sliced causing commands to
be split between two receiving events. This was causing the parser in
the AR9331 to submit incomplete commands to at ATMega32. The fix
consists in saving the partial command and to prepend it to the next
received buffer.