Fixed jerkiness in back to back move
The first move was longer by the length of a pulse causing a slight speed change between each movement. Made funcions called in the timer ISR inline to avoid ISRRAM for all of them. Got rid of the speed fixed point calculation and reorganized how a movement step count is performed